refactor(member): 활동 중복 지원 방지 및 activity queries index 파일 생성
