+++ apps/member/src/hooks/queries/activity/index.ts
@@ -9,3 +9,11 @@ export * from './useActivityGroupMember';
 export * from './useMyActivityGroupMember';
 export * from './useActivityPhoto';
 export * from './useActivityPhotoMutation';
+export * from './useActivityGroupMemberApplied';
+export * from './useActivityGroupMemberRoleMutation';
+export * from './useActivityGroupDeleteMutation';
+export * from './useActivityGroupStatusMutation';
+export * from './useActivityGroupAdminMutation';
+export * from './useActivityGroupBoardByParent';
+export * from './useActivityGroupMemberList';
+export * from './useActivityGroupMutation';
