+++ apps/member/src/hooks/queries/activity/useActivityGroupMemberMutation.ts
@@ -2,6 +2,7 @@ import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 import { postActivityGroupMemberApply } from '@api/activity';
 import { ACTIVITY_QUERY_KEY } from '@constants/key';
+import { API_ERROR_MESSAGE } from '@constants/message';
 import useToast from '@hooks/common/useToast';
 
 /**
@@ -13,18 +14,19 @@ export function useActivityGroupMemberMutation() {
 
   const mutation = useMutation({
     mutationFn: postActivityGroupMemberApply,
-    onSuccess: (data) => {
-      if (!data) {
+    onSuccess: ({ success, data: data, errorMessage }) => {
+      if (errorMessage) {
         toast({
           state: 'error',
-          message: '신청에 실패했습니다.',
+          message: API_ERROR_MESSAGE[errorMessage],
         });
-      } else {
+      } else if (success) {
         toast({
           state: 'success',
           message: '신청이 완료되었습니다.',
         });
       }
+
       queryClient.invalidateQueries({
         queryKey: ACTIVITY_QUERY_KEY.APPLICATION(data),
       });
